+++ lib/plugins/index.js
@@ -18,7 +18,7 @@ module.exports = [
   require('./plugin/plugin.js'),
   require('./plugin/install.js'),
   require('./plugin/uninstall/uninstall.js'),
-  require('./plugin/list/list.js'),
+  require('./plugin/list.js'),
   require('./plugin/search/search.js'),
   require('./aws/configCredentials.js'),
   require('./aws/provider.js'),
