+++ lib/plugins/plugin/list.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const BbPromise = require('bluebird');
-const pluginUtils = require('../lib/utils');
+const pluginUtils = require('./lib/utils');
 
 class PluginList {
   constructor(serverless, options) {
